+++ sketch_oxide/src/membership/mod.rs
@@ -17,6 +17,7 @@ mod scalable_bloom;
 mod stable_bloom;
 mod stacked_filter;
 mod vacuum_filter;
+mod vector_quotient_filter;
 
 pub use adaptive_quotient_filter::AdaptiveQuotientFilter;
 pub use binary_fuse::BinaryFuseFilter;
@@ -33,6 +34,7 @@ pub use scalable_bloom::ScalableBloomFilter;
 pub use stable_bloom::StableBloomFilter;
 pub use stacked_filter::StackedFilter;
 pub use vacuum_filter::{VacuumFilter, VacuumFilterStats};
+pub use vector_quotient_filter::VectorQuotientFilter;
 
 #[cfg(test)]
 mod tests {